Design shops, quests, and overhaul modules: (Arcadia) v0.3.0
This moderate update implements the framework for quests, shopping, and item generation. This also fixes a lot of bugs, redesigns the palette data storage, implements merit claiming, allows for the searching of individual merits, implements statistic pagination, implements casino replies that require specific criteria, tweaks the item framework, implements recipes, implements boosters, tweaks unique item data storage, implements multiplayer game session result updating, and implements item grouping.

+++ Orikivo/src/Desync/Generation/CatalogGenerator.cs
@@ -128,7 +128,7 @@ namespace Orikivo.Desync
                     {
                         int discountToApply = RandomProvider.Instance.Next(0, entry.MaxPossibleDiscount);
 
-                        float rawDiscount = discountToApply / 100;
+                        float rawDiscount = discountToApply / (float) 100;
 
                         discountEntries.Add(entry.ItemId, rawDiscount);
                         discounts++;
